analysis/requirements.txt
natsort>=6.0.0
scipy>=1.3.1
numpy>=1.17.0
opencv_python>=4.1.0.25
matplotlib>=3.1.1
Pillow>=6.1.0
